+++ case/stak_and_delegate/delegate/__init__.py
@@ -1,0 +1,1 @@
+# -*- coding: utf-8 -*-
