+++ case/stak_and_delegate/staking/ppos_params.json
@@ -1,0 +1,115 @@
+{
+	"P1000":{
+		"Typ":1,
+		"BenifitAddress":"0x12c171900f010b17e969702efa044d077e868082",
+		"NodeId": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
+		"ExternalId":"111111",
+		"NodeName": "platon",
+		"Website": "https://www.test.network",
+		"Details": "supper node",
+		"Amount":1000000000000000000000000,
+		"ProcessVersion":1792
+	},
+	"P1001":{
+		"BenifitAddress":"0x12c171900f010b17e969702efa044d077e868082",
+		"NodeId": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
+		"ExternalId":"111111",
+		"NodeName": "platon",
+		"Website": "https://www.test.network",
+		"Details": "supper node"
+	},
+	"P1002":{
+		"NodeId": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
+		"Typ":1,
+		"Amount":1000000000000000000000000
+	},
+	"P1003":{
+		"NodeId": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429"
+	},
+	"P1004":{
+		"Typ":1,
+		"NodeId": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
+		"Amount":1000000000000000000000000
+	},
+	"P1005":{
+		"StakingBlockNum":1000,
+		"NodeId": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
+		"Amount":1000000000000000000000000
+	},
+	"P1103":{
+		"Addr":"0x12c171900f010b17e969702efa044d077e868082"
+	},
+	"P1104":{
+		"StakingBlockNum":1000,
+		"DelAddr":"0x12c171900f010b17e969702efa044d077e868082",
+		"NodeId": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429"
+	},
+	"P1105":{
+		"NodeId": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429"
+	},
+	"P2000":{
+		"Verifier": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
+		"GithubID": "GithubID",
+		"Topic":"Topic",
+		"Desc":"Desc",
+		"Url":"http://www.test.inet",
+		"EndVotingBlock":1000
+	},
+	"P2001":{
+		"Verifier": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
+		"GithubID": "GithubID",
+		"Topic":"Topic",
+		"Desc":"Desc",
+		"Url":"http://www.test.inet",
+		"NewVersion":1,
+		"EndVotingBlock":1000,
+		"ActiveBlock":1000
+	},
+	"P2002":{
+		"Verifier": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
+		"GithubID": "GithubID",
+		"Topic":"Topic",
+		"Desc":"Desc",
+		"Url":"http://www.test.inet",
+		"ParamName":"ParamName",
+		"CurrentValue":"0.85",
+		"NewValue":"1.02",
+		"EndVotingBlock":1000
+	},
+	"P2003":{
+		"Verifier": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
+		"ProposalID": "0x12c171900f010b17e969702efa044d077e86808212c171900f010b17e969702e",
+		"Op":1
+	},
+	"P2004":{
+		"ActiveNode": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
+		"Version": 1
+	},
+	"P2100":{
+		"ProposalID": "0x12c171900f010b17e969702efa044d077e86808212c171900f010b17e969702e"
+	},
+	"P2101":{
+		"ProposalID": "0x12c171900f010b17e969702efa044d077e86808212c171900f010b17e969702e"
+	},
+	"P3000":{
+		"Data":"0x12c171900f010b17e969702efa044d077e868082"
+	},
+	"P3001":{
+		"Etype":1,
+		"Addr":"0x12c171900f010b17e969702efa044d077e868082",
+		"BlockNumber":1000
+	},
+	"P4000":{
+		"Account":"0x12c171900f010b17e969702efa044d077e868082",
+		"Plans":[{
+			"Epoch":5,
+			"Amount":1000000000000000000000000
+		},{
+			"Epoch":6,
+			"Amount":2000000000000000000000000
+		}]
+	},
+	"P4100":{
+		"Account":"0x12c171900f010b17e969702efa044d077e868082"
+	}
+}
