--- case/staking/ppos_params.json
@@ -1,13 +1,0 @@
-{
-	"P1000": {
-		"Typ": 1,
-		"BenifitAddress": "0x12c171900f010b17e969702efa044d077e868082",
-		"NodeId": "1f3a8672348ff6b789e416762ad53e69063138b8eb4d8780101658f24b2369f1a8e09499226b467d8bc0c4e03e1dc903df857eeb3c67733d21b6aaee2840e429",
-		"ExternalId": "111111",
-		"NodeName": "platon",
-		"Website": "https://www.test.network",
-		"Details": "supper node",
-		"Amount": 1000000000000000000000000,
-		"ProcessVersion": 1792
-	}
-}
